requirements.txt
Flask==2.0.1
google-cloud-storage==1.42.0
google-cloud-aiplatform==1.12.1

